Move map data hooks to makeData file
