+++ moped-editor/src/views/projects/projectView/ProjectComponents/utils/makeData.js
@@ -1,3 +1,6 @@
+import { useMemo } from "react";
+import { featureTableFieldMap } from "./makeFeatures";
+
 /**
  * Create an object map of component feature collections by component id
  * Ex. { 1: { type: "FeatureCollection", features: [...] } }
@@ -34,3 +37,42 @@ export const makeComponentFeatureCollectionsMap = (components) => {
 
   return componentGeographyMap;
 };
+
+// returns geojson of features across all components
+export const useProjectFeatures = (components) =>
+  useMemo(() => {
+    if (components === null || components.length === 0)
+      return {
+        type: "FeatureCollection",
+        features: [],
+      };
+
+    const allComponentfeatures = [];
+
+    components.forEach((component) => {
+      Object.keys(featureTableFieldMap).forEach((key) => {
+        if (component.hasOwnProperty(key))
+          allComponentfeatures.push(component[key]);
+      });
+    });
+
+    // Make features valid GeoJSON by adding type and properties attributes
+    const geoJsonFeatures = allComponentfeatures.flat().map((component) => ({
+      ...component,
+      type: "Feature",
+      properties: {},
+    }));
+
+    return {
+      type: "FeatureCollection",
+      features: geoJsonFeatures,
+    };
+  }, [components]);
+
+export const useDraftComponentFeatures = (draftComponent) =>
+  useMemo(() => {
+    return {
+      type: "FeatureCollection",
+      features: draftComponent?.features || [],
+    };
+  }, [draftComponent]);
